refactor muxrpc tests
